Fix warnigns

+++ app/src/main/java/io/bytesatwork/skycell/connectivity/CloudUploader.java
@@ -14,8 +14,8 @@ import android.util.Log;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FilenameFilter;
 import java.io.InputStreamReader;
+import java.util.Objects;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
@@ -30,9 +30,9 @@ import static java.util.concurrent.TimeUnit.SECONDS;
 public class CloudUploader {
     private static final String TAG = CloudUploader.class.getSimpleName();
 
-    private SkyCellApplication app;
-    private ScheduledExecutorService mExecutor;
-    private CloudConnection mConnection;
+    private final SkyCellApplication app;
+    private final ScheduledExecutorService mExecutor;
+    private final CloudConnection mConnection;
     private ScheduledFuture<?> mFuture;
 
     public CloudUploader() {
@@ -45,7 +45,7 @@ public class CloudUploader {
 
     public void start() {
         Log.i(TAG + ":" + Utils.getLineNumber(), "Start");
-        String path = app.getApplicationContext().getExternalFilesDir(null).getAbsolutePath();
+        String path = Objects.requireNonNull(app.getApplicationContext().getExternalFilesDir(null)).getAbsolutePath();
         String url = app.mSettings.loadString(Settings.SHARED_PREFERENCES_URL_UPLOAD);
         String apiKey = app.mSettings.loadString(Settings.SHARED_PREFERENCES_APIKEY);
         long rate = app.mSettings.loadLong(Settings.SHARED_PREFERENCES_UPLOAD_RATE_SECS);
@@ -87,35 +87,32 @@ public class CloudUploader {
             Log.i(TAG + ":" + Utils.getLineNumber(), "run");
             if (Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
                 if (mDirectory.isDirectory()) {
-                    File[] files = mDirectory.listFiles(new FilenameFilter() {
-                        @Override
-                        public boolean accept(File dir, String name) {
-                            return name.toLowerCase().endsWith(Constants.FILE_ENDING);
-                        }
-                    });
-                    for (int i = 0; i < files.length; ++i) {
-                        Log.i(TAG + ":" + Utils.getLineNumber(), "Read: (" +
-                            (i + 1) + " / " + files.length + ") " + files[i].getName());
-                        try {
-                            BufferedReader buffreader = new BufferedReader(new InputStreamReader(
-                                new FileInputStream(files[i])));
-                            StringBuffer json = new StringBuffer("");
-                            String line = "";
-                            while ((line = buffreader.readLine()) != null) {
-                                json.append(line);
-                            }
-                            buffreader.close();
-                            if (json.length() > 0) {
-                                if (mConnection.post(json.toString(), mUrl, mApikey) != null) {
-                                    Log.i(TAG + ":" + Utils.getLineNumber(),
-                                        "Upload ok - delete file");
-                                    files[i].delete();
-                                } else {
-                                    Log.w(TAG + ":" + Utils.getLineNumber(), "Upload failed");
+                    File[] files = mDirectory.listFiles((dir, name) -> name.toLowerCase().endsWith(Constants.FILE_ENDING));
+                    if (files != null) {
+                        for (int i = 0; i < files.length; ++i) {
+                            Log.i(TAG + ":" + Utils.getLineNumber(), "Read: (" +
+                                (i + 1) + " / " + files.length + ") " + files[i].getName());
+                            try {
+                                BufferedReader buffreader = new BufferedReader(new InputStreamReader(
+                                    new FileInputStream(files[i])));
+                                StringBuilder json = new StringBuilder();
+                                String line;
+                                while ((line = buffreader.readLine()) != null) {
+                                    json.append(line);
                                 }
+                                buffreader.close();
+                                if (json.length() > 0) {
+                                    if (mConnection.post(json.toString(), mUrl, mApikey) != null) {
+                                        boolean ok = files[i].delete();
+                                        Log.i(TAG + ":" + Utils.getLineNumber(),
+                                            "Upload ok - delete file: " + ok);
+                                    } else {
+                                        Log.w(TAG + ":" + Utils.getLineNumber(), "Upload failed");
+                                    }
+                                }
+                            } catch (Exception e) {
+                                e.printStackTrace();
                             }
-                        } catch (Exception e) {
-                            e.printStackTrace();
                         }
                     }
                 }
